First parent follows 'release-pygplates' branch but all release files obtained from second parent (commit 'b539427').
This is achieved with the following (equivalent to 'git merge -s theirs' if it was possible):
- git checkout release-pygplates  # Start on 'release-pygplates' branch.
- git merge -s ours <second_parent>  # Merge strategy 'ours' never fails.
- git checkout --detach <second_parent>  # Change content to <second_parent> and detach HEAD (don't attach to <second_parent>).
- git reset --soft release-pygplates  # Move HEAD to 'release-pygplates' leaving working-tree/index unchanged.
- git checkout release-pygplates  # Attach HEAD to 'release-pygplates' branch.
- git commit --amend -C HEAD  # Change merge commit content to <second_parent> retaining commit message.

+++ src/app-logic/ResolvedTriangulationUtils.h
@@ -32,6 +32,8 @@
 #include <utility>
 #include <vector>
 
+#include <boost/numeric/conversion/cast.hpp>
+
 #include <CGAL/Origin.h> // Must come before <CGAL/barycenter.h> since it uses it but doesn't define it.
 #include <CGAL/barycenter.h>
 #include <CGAL/centroid.h>
@@ -82,7 +84,9 @@ namespace GPlatesAppLogic
 				// Attempt to insert the vertex into the map.
 				std::pair<typename vertex_index_map_type::iterator, bool> vertex_insert_result =
 						d_vertex_index_map.insert(
-								typename vertex_index_map_type::value_type(vertex, d_vertices.size()/*index*/));
+								typename vertex_index_map_type::value_type(
+										vertex,
+										boost::numeric_cast<unsigned int>(d_vertices.size()/*index*/)));
 				if (vertex_insert_result.second)
 				{
 					// Insertion successful - first time seen vertex - add vertex to the sequence.
